+++ eslint.config.mjs
@@ -1,0 +1,23 @@
+import { globalIgnores } from 'eslint/config'
+import eslint from '@eslint/js'
+import tseslint from 'typescript-eslint'
+import stylistic from '@stylistic/eslint-plugin'
+
+/** @type {import('@typescript-eslint/utils').TSESLint.FlatConfig.ConfigFile} */
+export default [
+  eslint.configs.recommended,
+  ...tseslint.configs.recommended,
+  {
+    plugins: {
+      '@stylistic': stylistic
+    },
+
+    rules: {
+      'no-control-regex': 'off',
+      '@stylistic/quotes': ['error', 'single'],
+      '@stylistic/semi': ['error', 'never'],
+    },
+  },
+  globalIgnores(['dist/*'])
+]
+